third_party/WebKit/LayoutTests/external/wpt/webmidi/midi-allowed-by-feature-policy.https.sub.html

<!DOCTYPE html>
<body>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=resources/featurepolicytest.js></script>
<script>
'use strict';

promise_test(
    () => getFeatureTestPromise(),
    'Feature-Policy header {"' + getFeatureName() + '": ["*"]} allows the top-level document.');

async_test(t => {
  assert_available_in_iframe(t, '', '#OK');
}, 'Feature-Policy header {"' + getFeatureName() + '": ["*"]} allows same-origin iframes.');

async_test(t => {
  assert_available_in_iframe(t, 'https://{{domains[www]}}:{{ports[https][0]}}', '#OK');
}, 'Feature-Policy header {' + getFeatureName() + ': ["*"]} allows cross-origin iframes.');
</script>
</body>
